@@ -8,15 +8,17 @@
     "eject": "expo eject"
   },
   "dependencies": {
-    "expo": "35.0.0",
+    "eslint-config-satya164": "3.1.2",
+    "expo": "36.0.0",
     "moment": "2.24.0",
-    "react": "16.8.3",
-    "react-dom": "16.8.3",
-    "react-native": "https://github.com/expo/react-native/archive/sdk-35.0.0.tar.gz"
+    "react": "16.9.0",
+    "react-dom": "16.9.0",
+    "react-native": "https://github.com/expo/react-native/archive/sdk-36.0.1.tar.gz",
+    "react-native-month": "1.0.2"
   },
   "devDependencies": {
-    "babel-plugin-module-resolver": "3.2.0",
-    "babel-preset-expo": "7.1.0"
+    "babel-plugin-module-resolver": "4.0.0",
+    "babel-preset-expo": "8.0.0"
   },
   "resolutions": {
     "@babel/runtime": "7.3.1"
