@@ -1,6 +1,6 @@
 {
   "name": "react-native-calendario",
-  "version": "1.0.0",
+  "version": "2.0.0",
   "description": "React Native Calendar",
   "main": "lib/commonjs/index.js",
   "react-native": "lib/module/index.js",
@@ -32,43 +32,44 @@
   },
   "homepage": "https://github.com/maggialejandro/react-native-calendario#readme",
   "dependencies": {
-    "moment": "2.24.0"
+    "moment": "2.24.0",
+    "react-native-month": "1.0.2"
   },
   "peerDependencies": {
     "react": "*",
     "react-native": "*"
   },
   "devDependencies": {
-    "@babel/core": "7.7.4",
-    "@babel/runtime": "7.7.4",
-    "@react-native-community/bob": "0.7.0",
-    "@types/enzyme": "3.10.3",
-    "@types/jest": "24.0.24",
-    "@types/react": "16.8.14",
-    "@types/react-native": "0.57.46",
-    "@types/react-test-renderer": "16.8.3",
+    "@babel/core": "7.8.4",
+    "@babel/runtime": "7.8.4",
+    "@react-native-community/bob": "0.10.0",
+    "@types/enzyme": "3.10.5",
+    "@types/jest": "25.1.4",
+    "@types/react": "16.9.0",
+    "@types/react-native": "0.61.12",
+    "@types/react-test-renderer": "16.9.0",
     "@typescript-eslint/eslint-plugin": "2.8.0",
     "@typescript-eslint/parser": "2.12.0",
     "babel-core": "7.0.0-bridge.0",
-    "babel-eslint": "10.0.3",
-    "babel-jest": "24.9.0",
+    "babel-eslint": "10.1.0",
+    "babel-jest": "25.1.0",
     "babel-test": "0.2.3",
-    "enzyme": "3.10.0",
+    "enzyme": "3.11.0",
     "enzyme-adapter-react-16": "1.15.2",
-    "eslint": "6.7.0",
+    "eslint": "6.8.0",
     "eslint-config-satya164": "3.1.2",
-    "husky": "4.0.2",
-    "jest": "24.9.0",
+    "husky": "4.2.3",
+    "jest": "25.1.0",
     "jest-enzyme": "7.1.2",
-    "jsdom": "15.2.1",
-    "metro-react-native-babel-preset": "0.54.0",
+    "jsdom": "16.2.1",
+    "metro-react-native-babel-preset": "0.56.0",
     "prettier": "1.19.1",
-    "react": "16.8.3",
-    "react-dom": "16.8.3",
-    "react-native": "0.59.8",
-    "react-test-renderer": "16.8.3",
-    "ts-jest": "24.2.0",
-    "typescript": "3.7.2"
+    "react": "16.9.0",
+    "react-dom": "16.9.0",
+    "react-native": "0.61.5",
+    "react-test-renderer": "16.9.0",
+    "ts-jest": "25.2.1",
+    "typescript": "3.8.3"
   },
   "husky": {
     "hooks": {
