@@ -1,3 +1,2 @@
 export { default as Calendar } from './components/Calendar';
-
-export { CalendarType, DayType, ThemeType, LocaleType } from './types';
+export * from './types';
