@@ -13,45 +13,15 @@ export function isValidDate(date: Date): boolean {
   );
 }
 
-function leapYear(year: number): boolean {
-  return (year % 4 === 0 && year % 100 !== 0) || year % 400 === 0;
+export function isSameDate(one: Date, other: Date) {
+  return (
+    one.getDate() === other.getDate() &&
+    one.getMonth() === other.getMonth() &&
+    one.getFullYear() === other.getFullYear()
+  );
 }
 
-export function getNumberOfDaysInMonth(
-  monthIndex: number,
-  year: number
-): number {
-  switch (monthIndex) {
-    case 0:
-      return 31;
-    case 1:
-      return leapYear(year) ? 29 : 28;
-    case 2:
-      return 31;
-    case 3:
-      return 30;
-    case 4:
-      return 31;
-    case 5:
-      return 30;
-    case 6:
-      return 31;
-    case 7:
-      return 31;
-    case 8:
-      return 30;
-    case 9:
-      return 31;
-    case 10:
-      return 30;
-    case 11:
-      return 31;
-    default:
-      return 30;
-  }
-}
-
-export function getMonthNames(locale: LocaleType): string[] {
+export function getMonthNames(locale?: LocaleType): string[] {
   switch (locale) {
     case 'es':
       return [
@@ -129,33 +99,4 @@ export function getMonthNames(locale: LocaleType): string[] {
         'December',
       ];
   }
-}
-
-function getWeekdayNames(locale: LocaleType): string[] {
-  switch (locale) {
-    case 'es':
-      return ['Dom', 'Lun', 'Mar', 'Mié', 'Jue', 'Vie', 'Sab'];
-    case 'en':
-      return ['Sun', 'Mon', 'Tue', 'Wed', 'Thu', 'Fri', 'Sat'];
-    case 'fr':
-      return ['Dim', 'Lun', 'Mar', 'Mer', 'Jeu', 'Ven', 'Sam'];
-    case 'br':
-      return ['Dom', 'Seg', 'Ter', 'Qua', 'Qui', 'Sex', 'Sab'];
-    default:
-      return ['Sun', 'Mon', 'Tue', 'Wed', 'Thu', 'Fri', 'Sat'];
-  }
-}
-
-export function getDayNames(
-  locale: LocaleType,
-  firstDayMonday: boolean
-): string[] {
-  const days = getWeekdayNames(locale);
-  if (firstDayMonday) {
-    const sunday = days.shift() as string;
-    days.push(sunday);
-    return days;
-  }
-
-  return days;
 }
